+++ lib/app.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'home_page.dart';
+
+class App extends StatelessWidget {
+  const App({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: const HomePage(title: 'Flutter Demo Home Page'),
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+    );
+  }
+}
